login now independent of profile page
