+++ app/middleware/login.ts
@@ -1,0 +1,4 @@
+export default defineNuxtRouteMiddleware((to, _from) => {
+  const user = useUser();
+  if (!user.value) return navigateTo(`/login?redirect=${to.path}`);
+});
